Refactor IDataFormatter implementation
