+++ AccountantTool/ReoGrid/DataFormatter/DataFormatter.cs
@@ -1,0 +1,9 @@
+﻿using unvell.ReoGrid;
+
+namespace AccountantTool.ReoGrid.DataFormatter
+{
+    public abstract class DataFormatter
+    {
+        public abstract string FormatCell(Cell cell);
+    }
+}
