+++ AccountantTool/ReoGrid/DataFormatter/EmptyDataFormatter.cs
@@ -1,0 +1,9 @@
+﻿using unvell.ReoGrid;
+
+namespace AccountantTool.ReoGrid.DataFormatter
+{
+    public class EmptyDataFormatter : DataFormatter
+    {
+        public override string FormatCell(Cell cell) => null;
+    }
+}
